+++ src/infra/dev/eu-south-1/variables.tf
@@ -259,10 +259,14 @@ variable "sessions_table" {
 variable "client_registrations_table" {
   type = object({
     point_in_time_recovery_enabled = optional(bool, false)
+    stream_enabled                 = optional(bool, false)
+    stream_view_type               = optional(string, null)
   })
   description = "Client configurations table."
   default = {
     point_in_time_recovery_enabled = false
+    stream_enabled                 = true
+    stream_view_type               = "NEW_AND_OLD_IMAGES"
   }
 }
 
@@ -735,6 +739,50 @@ variable "api_alarms" {
       period              = 300
       statistic           = "Average"
       method              = "GET"
+      threshold           = 2000
+    },
+    "oidc-userinfo-get-5xx-error" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "5XXError"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Sum"
+      threshold           = 1
+      method              = "GET"
+    },
+    "oidc-userinfo-get-latency-alarm" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "Latency"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Average"
+      method              = "GET"
+      threshold           = 2000
+    },
+    "oidc-userinfo-post-5xx-error" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "5XXError"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Sum"
+      threshold           = 1
+      method              = "POST"
+    },
+    "oidc-userinfo-post-latency-alarm" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "Latency"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Average"
+      method              = "POST"
       threshold           = 2000
     },
     "oidc-register-5xx-error" = {