+++ src/infra/prod/eu-central-1/variables.tf
@@ -435,9 +435,9 @@ variable "ecs_alarms" {
     "cpu_high" = {
       metric_name         = "CPUUtilization"
       namespace           = "AWS/ECS"
-      evaluation_periods  = 1
+      evaluation_periods  = 2
       comparison_operator = "GreaterThanOrEqualToThreshold"
-      threshold           = 50
+      threshold           = 40
       period              = 60
       statistic           = "Average"
       scaling_policy      = "cpu_high"
@@ -635,6 +635,50 @@ variable "api_alarms" {
       period              = 300
       statistic           = "Average"
       method              = "GET"
+      threshold           = 1000
+    },
+    "oidc-userinfo-get-5xx-error" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "5XXError"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Sum"
+      threshold           = 1
+      method              = "GET"
+    },
+    "oidc-userinfo-get-latency-alarm" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "Latency"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Average"
+      method              = "GET"
+      threshold           = 1000
+    },
+    "oidc-userinfo-post-5xx-error" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "5XXError"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Sum"
+      threshold           = 1
+      method              = "POST"
+    },
+    "oidc-userinfo-post-latency-alarm" = {
+      resource_name       = "/oidc/userinfo"
+      metric_name         = "Latency"
+      namespace           = "AWS/ApiGateway"
+      evaluation_periods  = 2
+      comparison_operator = "GreaterThanOrEqualToThreshold"
+      period              = 300
+      statistic           = "Average"
+      method              = "POST"
       threshold           = 1000
     },
     "oidc-register-5xx-error" = {