+++ src/infra/uat/eu-south-1/main.tf
@@ -594,6 +594,26 @@ module "backend" {
     maximum_record_age_in_seconds = var.dlq_assertion_setting.maximum_record_age_in_seconds
   }
 
+  # TODO: enable cache updater lambda and pipe when the cache updater lambda and ecs core updates are ready to be deployed
+  # eventbridge_pipe_cache_updater = {
+  #   pipe_name                     = format("%s-cache-updater-pipe", local.project)
+  #   maximum_retry_attempts        = var.dlq_assertion_setting.maximum_retry_attempts
+  #   maximum_record_age_in_seconds = var.dlq_assertion_setting.maximum_record_age_in_seconds
+  # }
+
+  # cache_updater_lambda = {
+  #   name                              = format("%s-cache-updater", local.project)
+  #   filename                          = "${path.module}/../../hello-java/build/libs/hello-java-1.0-SNAPSHOT.jar"
+  #   cloudwatch_logs_retention_in_days = var.lambda_cloudwatch_logs_retention_in_days
+  #   vpc_id                            = module.network.vpc_id
+  #   vpc_subnet_ids                    = module.network.intra_subnets_ids
+  #   environment_variables = {
+  #     LOG_LEVEL              = var.app_log_level
+  #     CACHE_ENDPOINT_ADDRESS = module.client_cache.cache_endpoint_address
+  #     CACHE_ENDPOINT_PORT    = tostring(module.client_cache.cache_endpoint_port)
+  #   }
+  # }
+
   invalidate_cache_lambda = {
     name                             = format("%s-invalidate-cache", local.project)
     filename                         = "${path.module}/../../hello-python/lambda.zip"
@@ -682,6 +702,16 @@ module "backend" {
     namespace = "${local.project}-core/ApplicationMetrics"
   }
 }
+
+# TODO: enable cache updater lambda and pipe when the cache updater lambda and ecs core updates are ready to be deployed
+# module "client_cache" {
+#   source = "../../modules/cache"
+#   cache_name                 = format("%s-client-cache", local.project)
+#   vpc_id                     = module.network.vpc_id
+#   subnet_ids                 = module.network.private_subnet_ids
+#   allowed_security_group_ids = [module.backend.ecs_core_security_group_id, module.backend.cache_updater_security_group_id]
+#   alarm_sns_topic_arn        = module.sns.sns_topic_arn
+# }
 
 module "database" {
   source                      = "../../modules/database"