+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/exception/InvalidAccessTokenException.java
@@ -1,0 +1,10 @@
+package it.pagopa.oneid.exception;
+
+import com.nimbusds.oauth2.sdk.token.BearerTokenError;
+
+public class InvalidAccessTokenException extends RuntimeException {
+
+  public InvalidAccessTokenException() {
+    super(BearerTokenError.INVALID_TOKEN.getCode());
+  }
+}