+++ src/oneid/oneid-lambda-cache-updater/src/main/java/it/pagopa/oneid/service/DynamoStreamService.java
@@ -1,0 +1,14 @@
+package it.pagopa.oneid.service;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import it.pagopa.oneid.common.model.Client;
+import java.util.Optional;
+
+public interface DynamoStreamService {
+
+  Optional<Client> extractClient(JsonNode streamRecord, boolean preferOldImage);
+
+  Optional<String> extractClientId(JsonNode streamRecord, boolean preferOldImage);
+
+  boolean hasCacheRelevantChanges(JsonNode streamRecord);
+}